+++ fetchers/enrichers/__init__.py
@@ -1,6 +1,5 @@
 """Enrichers package - Event data enrichment using GPT and other services."""
 
-from .conference_gpt_extractor import enrich_conference_data
-from .hackathon_gpt_extractor import enrich_hackathon_data
+from .unified_gpt_extractor import enrich_conference_data, enrich_hackathon_data
 
 __all__ = ['enrich_conference_data', 'enrich_hackathon_data']
